+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add support 'late' field modifier.
 * Add support for `Expression.nullChecked` to add a null assertion operator.
 * Add support for creating `mixin`s.
+* Add `Expression.nullSafeSpread` for the null aware spread operator `...?`.
 
 ## 4.0.0